+++ breeze-core/src/main/kotlin/domain/MatchType.kt
@@ -1,7 +1,7 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
-package com.windea.breezeframework.core.domain.text
+package com.windea.breezeframework.core.domain
 
 import com.windea.breezeframework.core.extensions.*